requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    scripts/gen_requirements.sh
#
--extra-index-url https://us-python.pkg.dev/ce-builder/python/simple/

anyio==3.7.0
    # via httpcore
authlib==1.2.0
    # via spatiafi (setup.cfg)
autoflake==2.1.1
    # via spatiafi (setup.cfg)
black==23.3.0
    # via spatiafi (setup.cfg)
bleach==6.0.0
    # via readme-renderer
build==0.10.0
    # via
    #   pip-tools
    #   spatiafi (setup.cfg)
cachetools==5.3.1
    # via
    #   google-auth
    #   tox
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
chardet==5.1.0
    # via tox
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
    #   spatiafi (setup.cfg)
colorama==0.4.6
    # via tox
coverage==7.2.7
    # via pytest-cov
cryptography==41.0.1
    # via
    #   authlib
    #   secretstorage
distlib==0.3.6
    # via virtualenv
docutils==0.20.1
    # via readme-renderer
exceptiongroup==1.1.1
    # via
    #   anyio
    #   pytest
filelock==3.12.1
    # via
    #   tox
    #   virtualenv
flake8==6.0.0
    # via
    #   flake8-pyproject
    #   spatiafi (setup.cfg)
flake8-pyproject==1.2.3
    # via spatiafi (setup.cfg)
google-auth==2.19.1
    # via keyrings-google-artifactregistry-auth
h11==0.14.0
    # via httpcore
httpcore==0.17.2
    # via httpx
httpx==0.24.1
    # via spatiafi (setup.cfg)
identify==2.5.24
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==6.6.0
    # via
    #   keyring
    #   twine
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via spatiafi (setup.cfg)
jaraco-classes==3.2.3
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
keyring==23.13.1
    # via
    #   keyrings-google-artifactregistry-auth
    #   spatiafi (setup.cfg)
    #   twine
keyrings-google-artifactregistry-auth==1.1.2
    # via spatiafi (setup.cfg)
markdown-it-py==3.0.0
    # via rich
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
more-itertools==9.1.0
    # via jaraco-classes
mypy==1.3.0
    # via spatiafi (setup.cfg)
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.8.0
    # via pre-commit
packaging==23.1
    # via
    #   black
    #   build
    #   pyproject-api
    #   pytest
    #   setuptools-scm
    #   tox
pathspec==0.11.1
    # via black
pip-tools==6.13.0
    # via spatiafi (setup.cfg)
pkginfo==1.9.6
    # via twine
platformdirs==3.5.3
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.0.0
    # via
    #   keyrings-google-artifactregistry-auth
    #   pytest
    #   tox
pre-commit==3.3.2
    # via spatiafi (setup.cfg)
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via
    #   autoflake
    #   flake8
pygments==2.15.1
    # via
    #   readme-renderer
    #   rich
pyproject-api==1.5.1
    # via tox
pyproject-hooks==1.0.0
    # via build
pytest==7.3.2
    # via
    #   pytest-cov
    #   spatiafi (setup.cfg)
pytest-cov==4.1.0
    # via spatiafi (setup.cfg)
pyyaml==6.0
    # via pre-commit
readme-renderer==37.3
    # via twine
requests==2.31.0
    # via
    #   keyrings-google-artifactregistry-auth
    #   requests-toolbelt
    #   spatiafi (setup.cfg)
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.4.2
    # via twine
rsa==4.9
    # via google-auth
secretstorage==3.3.3
    # via keyring
setuptools-scm==7.1.0
    # via spatiafi (setup.cfg)
six==1.16.0
    # via
    #   bleach
    #   google-auth
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tomli==2.0.1
    # via
    #   autoflake
    #   black
    #   build
    #   coverage
    #   flake8-pyproject
    #   mypy
    #   pyproject-api
    #   pyproject-hooks
    #   pytest
    #   setuptools-scm
    #   tox
tox==4.6.0
    # via spatiafi (setup.cfg)
twine==4.0.2
    # via spatiafi (setup.cfg)
typing-extensions==4.6.3
    # via
    #   mypy
    #   setuptools-scm
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
    #   twine
virtualenv==20.23.0
    # via
    #   pre-commit
    #   tox
webencodings==0.5.1
    # via bleach
wheel==0.40.0
    # via
    #   pip-tools
    #   spatiafi (setup.cfg)
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
pip==23.1.2
    # via pip-tools
setuptools==67.8.0
    # via
    #   nodeenv
    #   pip-tools
    #   setuptools-scm
